Redo started, AMAZING results so far
This game so far feels REALLY satisfying to play. There isn't much there yet, but I think the general feel of the game is already nailed down pretty well.

Character controller is in, basic weapon functionality, animator is started, IK for player aim is about where it needs to be (though will need some adjusting for final character models)

Still lots to do, but I'm happy with progress so far.